surface: Use the unconfined apparmor profile for ghostty children
When launching any process inside ghostty it will use the ghostty
apparmor profile, this is problematic because any app will be considered
as part of such security environment, and GNOME Shell will use this to
map the windows as part of the same ghostty application.

To prevent this, ensure that the ghostty children will be launched in
the unconfined profile, as it would happen for any app.

Of course when launching another snap, it will be then added to the
expected apparmor profile.

So use apparmor APIs to decuple the children from the terminal instead

With this we've now:

  tricky ~
  ❯ cat /proc/self/attr/current
  unconfined

  tricky ~
  ❯ cat /proc/$(pidof ghostty)/attr/current
  snap.ghostty.ghostty (complain)